float instead double type, add time measurements of multiplication calculation

+++ kernels/naive_mult.cl
@@ -1,13 +1,13 @@
-__kernel void multiply(const __global double* A,
-                       const __global double* B,
-                       __global double* C,
+__kernel void multiply(const __global float* A,
+                       const __global float* B,
+                       __global float* C,
                        const int widthA, 
                        const int widthB) 
 {
     const int x = get_global_id(0);
     const int y = get_global_id(1);
  
-    double result = 0.0;
+    float result = 0.0;
     for (int i = 0; i < widthA; ++i) 
     {
         result += A[widthA * y + i] * B[widthB * i + x];
